+++ rest-dynamodb-modular/app/events/event-item-create.json
