+++ rest-dynamodb-modular/app/events/event-item-delete.json
