+++ rest-dynamodb-modular/app/events/event-item-find.json
