+++ rest-dynamodb-modular/app/events/event-item-list.json
