+++ rest-dynamodb-modular/app/events/event-item-update.json
