+++ rest-dynamodb-modular/app/src/__tests__/unit/handlers/item-delete.test.js
@@ -1,10 +1,10 @@
-// Import all functions from create-item.js
-const lambda = require('../../../handlers/create-item');
+// Import all functions from delete-item.js
+const lambda = require('../../../handlers/item-delete');
 // Import dynamodb from aws-sdk
 const { DynamoDBDocumentClient } = require('@aws-sdk/lib-dynamodb');
 
-// This includes all tests for CreateItem handler
-describe('handler::CreateItem', function () {
+// This includes all tests for FindItem handler
+describe('handler::DeleteItem', () => {
   let sendSpy;
 
   // Test one-time setup and teardown, see more in https://jestjs.io/docs/en/setup-teardown
@@ -24,24 +24,29 @@ describe('handler::CreateItem', function () {
     sendSpy.mockRestore();
   });
 
-  it('should create an item', async () => {
-    const returnedItem = { id: 'id1', name: 'name1' };
+  it('should return status code 204 when successful', async () => {
+    const item = { id: 'id1' };
 
-    // Return the specified value whenever the spied put function is called
-    sendSpy.mockResolvedValue(returnedItem);
+    // Return the specified value whenever the spied function is called
+    sendSpy.mockResolvedValue({});
 
     const event = {
-      httpMethod: 'POST',
-      body: '{"name": "name1"}',
+      httpMethod: 'DELETE',
+      pathParameters: {
+        itemId: 'id1',
+      },
     };
 
     // Invoke the handler
     const result = await lambda.handle(event);
 
+    const expectedResult = {
+      statusCode: 204,
+    };
+
     // Expect dynamodb to have been called
     expect(sendSpy).toHaveBeenCalledTimes(1);
     // Compare the result with the expected result
-    expect(result.body).toMatch(/name1/);
-    expect(result.statusCode).toEqual(201);
+    expect(result).toEqual(expectedResult);
   });
 });
