+++ rest-dynamodb-modular/app/src/__tests__/unit/handlers/item-update.test.js
@@ -1,10 +1,11 @@
-// Import all functions from find-item.js
-const lambda = require('../../../handlers/find-item');
+// Import all functions from update-item.js
+const lambda = require('../../../handlers/item-update');
 // Import dynamodb from aws-sdk
 const { DynamoDBDocumentClient } = require('@aws-sdk/lib-dynamodb');
+const { ConditionalCheckFailedException } = require('@aws-sdk/client-dynamodb');
 
-// This includes all tests for FindItem handler
-describe('handler::FindItem', () => {
+// This includes all tests for CreateItem handler
+describe('handler::UpdateItem', function () {
   let sendSpy;
 
   // Test one-time setup and teardown, see more in https://jestjs.io/docs/en/setup-teardown
@@ -24,56 +25,51 @@ describe('handler::FindItem', () => {
     sendSpy.mockRestore();
   });
 
-  it('should return item when found', async () => {
-    const item = { id: 'id1' };
+  it('should update an item', async () => {
+    const returnedItem = { Attributes: { id: 'id1', name: 'name1' } };
 
-    // Return the specified value whenever the spied function is called
-    sendSpy.mockResolvedValue({ Item: item });
+    // Return the specified value whenever the spied put function is called
+    sendSpy.mockResolvedValue(returnedItem);
 
     const event = {
-      httpMethod: 'GET',
+      httpMethod: 'PUT',
       pathParameters: {
         itemId: 'id1',
       },
+      body: '{"name": "name1"}',
     };
 
     // Invoke the handler
     const result = await lambda.handle(event);
 
-    const expectedResult = {
-      statusCode: 200,
-      body: JSON.stringify(item),
-    };
-
     // Expect dynamodb to have been called
     expect(sendSpy).toHaveBeenCalledTimes(1);
     // Compare the result with the expected result
-    expect(result).toEqual(expectedResult);
+    expect(result.body).toMatch(/name1/);
+    expect(result.statusCode).toEqual(200);
   });
 
   it('should return status code 404 when not found', async () => {
-    const item = { id: 'id1' };
+    const returnedItem = { Attributes: { id: 'id1', name: 'name1' } };
 
-    // Return the specified value whenever the spied function is called
-    sendSpy.mockReturnValue(Promise.resolve({}));
+    // Return the specified value whenever the spied put function is called
+    sendSpy.mockRejectedValue(new ConditionalCheckFailedException());
 
     const event = {
-      httpMethod: 'GET',
+      httpMethod: 'PUT',
       pathParameters: {
         itemId: 'id1',
       },
+      body: '{"name": "name1"}',
     };
 
     // Invoke the handler
     const result = await lambda.handle(event);
 
-    const expectedResult = {
-      statusCode: 404,
-    };
-
     // Expect dynamodb to have been called
     expect(sendSpy).toHaveBeenCalledTimes(1);
     // Compare the result with the expected result
-    expect(result).toEqual(expectedResult);
+    expect(result.body).toBeUndefined();
+    expect(result.statusCode).toEqual(404);
   });
 });
