+++ rest-dynamodb-modular/app/src/handlers/item-create.js
