+++ rest-dynamodb-modular/app/src/handlers/item-delete.js
