+++ rest-dynamodb-modular/app/src/handlers/item-list.js
