+++ rest-dynamodb-modular/app/src/handlers/item-update.js
