+++ rest-dynamodb-modular/template.yaml
@@ -1,9 +1,9 @@
 AWSTemplateFormatVersion: '2010-09-09'
 Transform: AWS::Serverless-2016-10-31
 Description: >
-  rest-dynamodb
+  rest-dynamodb-modular
 
-  SAM Template for rest-dynamodb
+  SAM Template for rest-dynamodb-modular
 
 Globals:
   Function:
@@ -19,7 +19,7 @@ Resources:
   listItemsFunction:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: src/handlers/list-items.handle
+      Handler: src/handlers/item-list.handle
       Description: A simple example includes a HTTP get method to get all items from a DynamoDB table.
       Policies:
         # Give Create/Read/Update/Delete Permissions to the SampleTable
@@ -35,7 +35,7 @@ Resources:
   findItemFunction:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: src/handlers/find-item.handle
+      Handler: src/handlers/item-find.handle
       Description: A simple example includes a HTTP get method to get one item by id from a DynamoDB table.
       Policies:
         # Give Create/Read/Update/Delete Permissions to the SampleTable
@@ -51,7 +51,7 @@ Resources:
   createItemFunction:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: src/handlers/create-item.handle
+      Handler: src/handlers/item-create.handle
       Description: A simple example includes a HTTP post method to add one item to a DynamoDB table.
       Policies:
         # Give Create/Read/Update/Delete Permissions to the SampleTable
@@ -67,7 +67,7 @@ Resources:
   updateItemFunction:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: src/handlers/update-item.handle
+      Handler: src/handlers/item-update.handle
       Description: A simple example includes a HTTP put method to update one item by id in a DynamoDB table.
       Policies:
         # Give Create/Read/Update/Delete Permissions to the SampleTable
@@ -83,7 +83,7 @@ Resources:
   deleteItemFunction:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: src/handlers/delete-item.handle
+      Handler: src/handlers/item-delete.handle
       Description: A simple example includes a HTTP delete method to delete one item by id from a DynamoDB table.
       Policies:
         # Give Create/Read/Update/Delete Permissions to the SampleTable
